+++ GamesAPI/GamesAPI.Models/ResultMessageModel.cs
@@ -8,7 +8,7 @@ namespace GamesAPI.Models
     public class ResultMessageModel
     {
         public int Result { get; set; }
-        public int ResultID { get; set; }
+        public long ResultID { get; set; }
         public string ResultMessage { get; set; }
     }
 }
